feat: refactor, permit ID override, updated output method, additional metadata output
Refactored main function to reduce the amount of code repetition.

Added a check for filename ID and file ID to see if they differ; if they do, change the file ID to filename ID. This allows for manual renaming of ab1 files and for those changes to be reflected in the fasta files.

SeqIO.write method is now solely used for output of the fasta files.

Metadata output now includes the trace score and median PUP scores to permit ease of analysis. Additionally, the metadata rows are sorted by primer, then by sample name.